requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
billiard==3.6.4.0
black==21.5b1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
factory-boy==3.2.0
Faker==8.2.0
flake8==3.9.2
idna==2.10
kombu==5.0.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
prompt-toolkit==3.0.18
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
